refactor: splited 'parser' into a separate package
