+++ parser/lexer.go
@@ -1,5 +1,5 @@
 //go:generate stringer -type lexstate
-package cofunc
+package parser
 
 import (
 	"fmt"
@@ -98,7 +98,7 @@ func (l *lexer) split(line string, ln int) error {
 				l._goto(_lx_var_directuse1)
 				break
 			}
-			return ParseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
+			return parseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
 		case _lx_symbol:
 			if is.Symbol(c) {
 				l.saveRune(c)
@@ -137,7 +137,7 @@ func (l *lexer) split(line string, ln int) error {
 				l._goto(_lx_var_directuse1)
 				break
 			}
-			return ParseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
+			return parseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
 		case _lx_ident:
 			if is.Ident(c) {
 				l.saveRune(c)
@@ -163,7 +163,7 @@ func (l *lexer) split(line string, ln int) error {
 				l._goto(_lx_symbol)
 				break
 			}
-			return ParseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
+			return parseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
 		case _lx_string:
 			if is.BackSlash(c) {
 				l._goto(_lx_string_backslash)
@@ -191,7 +191,7 @@ func (l *lexer) split(line string, ln int) error {
 				l._goto(_lx_var_directuse2)
 				break
 			}
-			return ParseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
+			return parseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
 		case _lx_var_directuse2:
 			if is.Ident(c) {
 				l.saveRune(c)
@@ -206,7 +206,7 @@ func (l *lexer) split(line string, ln int) error {
 				l._goto(_lx_unknow)
 				break
 			}
-			return ParseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
+			return parseErrorf(ln, ErrTokenCharacterIllegal, "character '%c', state '%s'", c, l.state)
 		}
 	}
 	return nil
