+++ parser/token_test.go
@@ -1,4 +1,4 @@
-package cofunc
+package parser
 
 import (
 	"testing"
